MODSIDECAR-140: Use SECURE_STORE_ENV, not ENV, for secure store key
https://folio-org.atlassian.net/browse/MODSIDECAR-140

Summary:

Use the SECURE_STORE_ENV environment variable, not the ENV environment variable, as the first component of the secure store key.

Description:

It’s a well known pattern to run a single Kafka cluster for multiple flower releases: https://github.com/folio-org/rfcs/blob/master/text/0002-kafka-tenant-collection-topics.md#migration-to-new-module-version

The ENV environment variable is used as the first component of the Kafka topic. Setting ENV to the flower name (like ramsons or sunflower) ensures that topics are produced and consumed by modules of the same flower release only.

Using ENV for the secure store key is wrong because a password doesn’t change when migrating a tenant from ramsons to sunflower.

There can be a need to separate the entries in the secure store, for example using prod and test, but not by using the flower names from the ENV environment variable.

Fix, see https://folio-org.atlassian.net/browse/FOLIO-4344 :

* replace ENV by SECURE_STORE_ENV environment variable if used for secure store key
* fall back to ENV if SECURE_STORE_ENV is not defined to ensure upgrade compatibility for Ramsons CSP and Sunflower CSP; this fall back is not used for Trillium
* fall back to folio if neither SECURE_STORE_ENV nor ENV are defined (the same as the ENV default to ensure compatibility)
* module’s README explains SECURE_STORE_ENV